change float to double

+++ exercises/q1.hs
@@ -1,3 +1,3 @@
 -- Funcao que recebe o salario de um funcionario e retorna o salario com 3% de aumento
-salario :: Float -> Float
+salario :: Double -> Double
 salario n = n * (103 / 100)
